add in pages: register: home, register, settings
