+++ src/components/App.tsx
@@ -1,14 +1,14 @@
 import Header from "./header";
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 import { Route, Switch } from "react-router-dom";
 import Article from "./Article";
 import Editor from "./Editor";
-import Home from "./Home";
+import Home from "../pages/home-page/home-page";
 import Login from "./login";
 import Profile from "./Profile";
-import Register from "./register";
-import Settings from "./Settings";
+import Register from "../pages/register/register";
+import Settings from "../pages/settings/settings";
 import { store } from "../services/store";
 import { push } from "react-router-redux";
 import { auth, setTokenAxios } from "../api";
